minor fixes

+++ BlueAlienship.cs
@@ -30,8 +30,8 @@ namespace Space_Shooter
             if (Y >= 0 && Y <= Global.Height / 2) _gunSystem.AutoFire(X, Y);
             _gunSystem.Update();
             _movePattern.Update();
-            Y = (int)_movePattern.UpdatedY;
-            X = (int)_movePattern.UpdatedX;
+            Y = _movePattern.UpdatedY;
+            X = _movePattern.UpdatedX;
         }
     }
 }
